+++ src/texture/texture_object.cpp
@@ -27,7 +27,7 @@ bool TextureObject::load(String filepath)
 		return false;
 	}
 
-	const size_t fileSize = file->size();
+	const size_t fileSize = static_cast<size_t>(file->size());
 	UniquePtr<uint8_t[]> buffer(new uint8_t[fileSize]);
 	if (!file->blockRead((char *)buffer.get(), 0, fileSize))
 	{
@@ -92,7 +92,7 @@ bool TextureObject::loadDDS(String filepath)
 	}
 	else
 	{
-		const size_t fileSize = file->size();
+		const size_t fileSize = static_cast<size_t>(file->size());
 		UniquePtr<uint8_t[]> buffer(new uint8_t[fileSize]);
 		file->read((char *)buffer.get(), sizeof(char), fileSize);
 		file.reset();
